requirements.txt
streamlit>=1.32
Pillow>=10.0
numpy>=1.24
web3>=6.0
transformers==4.46.3
torch==2.5.1
torchvision==0.20.1
pandas>=2.0
